Add home screen with contact list, search, sort, and urgency-indicating list items.
