+++ lib/presentation/screens/home_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:intl/intl.dart';
 
 import 'package:contact_assistant/logic/providers/contact_provider.dart';
+import 'package:contact_assistant/presentation/widgets/contact_list_item.dart';
 
 class HomeScreen extends ConsumerStatefulWidget {
   const HomeScreen({super.key});
@@ -20,42 +20,6 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
   void dispose() {
     _searchController.dispose();
     super.dispose();
-  }
-
-  Color _getUrgencyColor(DateTime? lastContacted) {
-    if (lastContacted == null) return Colors.grey.shade300;
-    final difference = DateTime.now().difference(lastContacted).inDays;
-    if (difference < 14) {
-      return Colors.green.shade100;
-    } else if (difference < 30) {
-      return Colors.yellow.shade100;
-    } else {
-      return Colors.red.shade100;
-    }
-  }
-
-  IconData _getUrgencyIcon(DateTime? lastContacted) {
-    if (lastContacted == null) return Icons.question_mark;
-    final difference = DateTime.now().difference(lastContacted).inDays;
-    if (difference < 14) {
-      return Icons.check_circle;
-    } else if (difference < 30) {
-      return Icons.warning;
-    } else {
-      return Icons.error;
-    }
-  }
-
-  Color _getUrgencyIconColor(DateTime? lastContacted) {
-    if (lastContacted == null) return Colors.grey;
-    final difference = DateTime.now().difference(lastContacted).inDays;
-    if (difference < 14) {
-      return Colors.green;
-    } else if (difference < 30) {
-      return Colors.orange;
-    } else {
-      return Colors.red;
-    }
   }
 
   @override
@@ -127,35 +91,7 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
                   itemCount: filteredContacts.length,
                   itemBuilder: (context, index) {
                     final contact = filteredContacts[index];
-                    return Card(
-                      color: _getUrgencyColor(contact.lastContacted),
-                      margin: const EdgeInsets.symmetric(
-                          horizontal: 8, vertical: 4),
-                      child: ListTile(
-                        leading: CircleAvatar(
-                          backgroundColor: Colors.white,
-                          child: Text(contact.name.isNotEmpty
-                              ? contact.name[0].toUpperCase()
-                              : '?'),
-                        ),
-                        title: Text(contact.name,
-                            style:
-                                const TextStyle(fontWeight: FontWeight.bold)),
-                        subtitle: Text(
-                          contact.lastContacted == null
-                              ? 'Never contacted'
-                              : 'Last: ${DateFormat.yMMMd().format(contact.lastContacted!)}',
-                        ),
-                        trailing: Icon(
-                          _getUrgencyIcon(contact.lastContacted),
-                          color: _getUrgencyIconColor(contact.lastContacted),
-                        ),
-                        onTap: () {
-                          context.push('/contact/${contact.id}',
-                              extra: contact);
-                        },
-                      ),
-                    );
+                    return ContactListItem(contact: contact);
                   },
                 );
               },
